\name{MURAT-internal} 
\alias{Para_Match_Kur} 
\alias{Optimal_PValue} 
\alias{Int_Fnc} 
\alias{gls_sun} 
\alias{Eta_RV} 
\title{Internal MURAT Functions} 
\description{ 
  Internal MURAT functions 
}      
\details{ 
          These are not to be called by the user. 
} 
\keyword{ internal } 