\name{sample}
\alias{sample}
\docType{data}
\title{
Example Data 
}
\description{
Example data for testing the main function.
}
\usage{data("sample")}
\format{
List of 3

 $ G: a genotype matrix with rows for individuals and columns for SNPs.
 
 $ X: a covariate matrix with rows for individuals and columns for covariates, where the first column represents for intercept.
 
 $ Y: a phenotype matrix with rows for individuals and columns for multiple continuous traits.
}

%\details{
%To be added to deleted
%}
%\source{
%To be added or removed.
%}
%\references{
%To be added.
%}
\examples{
data(sample)
}
\keyword{dataset}
